refactor: move selector methods to CommandUtils
